Wire up the text section editor
I can now edit a card section... That's very exciting. It just works.

I mean, I guess I would not be anywhere near this surprised, and I would
have got this working much faster, if I was using react or something. So
I'm not sure this all makes practical sense in terms of optimising a
project for actual, like, output, but I am enjoying this.