Add basic tags editor tags implementation
And also do some standardisation of all of these methods for controlling
lists of things on custom elements. They should now all error when
trying to add an index that already exists. I think this is important,
as otherwise the DOM state could easily get out of sync with the state
that the component is tracking.

I wonder how straight forward it would be to abstract out these adders
and removers. I think for now that would probably still be a hasty
abstraction.