Make the card controller put initial card in view
It would previously require that all the properties on the card were set
after the controller was wired up.

This change is necessary to faciliate a controller ever changing cards,
but also just brings it inline with how I wrote the editor, and makes it
easier to use in code.