Update models for headings and back sections
This is entirely untested, and will absolutely break the UI. I guess
this is the first use of the multiple-versions-of-serialisation
"migration" stuff though. I could have just changed things in place,
given that I'm the only one that has ever run this, and hence the only
one for which this is a breaking change, but I think it's nice to try it
out, and to have an example of it for when I need to do it properly in
the future.