Get the app working again
No support for interacting with the back-sections in the UI yet, I have
just got it back working for the front-sections.

As headings are no longer their own thing, and are just sections, most
of the work was in wiring up the heading editor and card components.

This does actually mean there is some new functionality here, as you can
now have headings and subheadings wherever you want, and you can choose
to have them left and right aligned, as well as centred.